pyproject.toml
[project]
name = "jDas"
version = "0.1.0"
requires-python = ">=3.9"
description = "DAS data cleaning service"
dependencies = [
    "torch>=2.0",
    "pytorch-lightning>=2.0",
    "fastapi>=0.104",
    "uvicorn>=0.24",
    "hydra-core>=1.3",
    "omegaconf>=2.3",
    "obspy>=1.4",
    "scipy>=1.11",
    "numpy>=1.24",
    "pandas>=2.0",
    "requests>=2.31",
    "python-dotenv>=1.0",
    "dvc>=3.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1",
    "pre-commit>=3.5",
    "jupyter>=1.0",
    "matplotlib>=3.8",
    "python-dotenv>=1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["jDas"]

[project.scripts]
jDas = "jDas.cli.main:cli"

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B"]

[tool.ruff.format]
quote-style = "double"
